refactor(doctor): split preflight helpers
